requirements.txt

django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
python-dotenv==1.0.0
pillow==10.1.0
django-filter==23.3
